Add SensorResponseCity base class
